Removed original statistics layout
