+++ app/assets/javascripts/Components/assignment_summary.jsx
@@ -1,0 +1,45 @@
+import React from "react";
+import {render} from "react-dom";
+import {AssignmentSummaryTable} from "./assignment_summary_table";
+import {AssignmentChart} from "./assignment_chart";
+import {Tab, Tabs, TabList, TabPanel} from "react-tabs";
+
+class AssignmentSummary extends React.Component {
+  render() {
+    if (this.props.is_instructor) {
+      return (
+        <Tabs>
+          <TabList>
+            <Tab>{I18n.t("summary_table")}</Tab>
+            <Tab>{I18n.t("summary_statistics")}</Tab>
+          </TabList>
+          <TabPanel>
+            <AssignmentSummaryTable
+              course_id={this.props.course_id}
+              assignment_id={this.props.assessment_id}
+              is_instructor={this.props.is_instructor}
+            />
+          </TabPanel>
+          <TabPanel>
+            <AssignmentChart
+              course_id={this.props.course_id}
+              assessment_id={this.props.assessment_id}
+            />
+          </TabPanel>
+        </Tabs>
+      );
+    } else {
+      return (
+        <AssignmentSummaryTable
+          course_id={this.props.course_id}
+          assignment_id={this.props.assessment_id}
+          is_instructor={this.props.is_instructor}
+        />
+      );
+    }
+  }
+}
+
+export function makeAssignmentSummary(elem, props) {
+  return render(<AssignmentSummary {...props} />, elem);
+}
